create user todo

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServicer } from '@/auth/actions/auth-actions'
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup'
@@ -26,11 +27,19 @@ const postSchema = yup.object({
 })
 
 export async function POST(request: Request) {
+  const user = await getUserSessionServicer()
+
+  if (!user) {
+    return NextResponse.json('No autorized', { status: 401 })
+  }
+
   try {
     const { complete, description } = await postSchema.validate(
       await request.json()
     )
-    const todo = await prisma.todo.create({ data: { complete, description } })
+    const todo = await prisma.todo.create({
+      data: { complete, description, userId: user.id }
+    })
 
     return NextResponse.json(todo)
   } catch (error) {
@@ -39,8 +48,14 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getUserSessionServicer()
+
+  if (!user) {
+    return NextResponse.json('No autorized', { status: 401 })
+  }
+
   try {
-    await prisma.todo.deleteMany({ where: { complete: true } })
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } })
 
     return NextResponse.json('Deleted completed todos')
   } catch (error) {
